requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-community
langchain-groq
